Updated PortfolioOverview/Commands, using hooks to clean up
